LPS-128276 Can we inline this?
